Change in how nodes are moved, and how tactic is specified
Nodes are now only moved after an iteration. As a result, only as many nodes
can be moved as empty nodes. This results in changes to movement_tactics

Movement tactics are now specified in simulation_parameters, as a type
The class is created in schelling_model so the constructor is called
at the appropriate time