Probability based simulation improvements
End of infection, and end of recovery immunity are now modeled as erlang
distributions. Distance moved is now modeled as a normal distribution.
Values are sampled from functions in simulation_parameters. End of infection
and end of recovery are now simulated hourly

Mortality rate is now implemented as chance to die after end of infection