Major overhaul to support tkinter GUI
GUI is now launched from gui.py and this is the primary way of running the
simulation.
Simulation parameters now in a string-keyed dictionary to allow for easy
feeding of parameters from tkinter.
As a result, sanity checks are now in a function, and also perform type checking.
Static and dynamic runners are now inside functions, called from gui.py.